refactor: partially refactor iter types into submodules
